requirements.txt
google-genai 
youtube_transcript_api
